Move curses UI to separate class
